Combine pacu.py functions and util class into Main
Given that most of the functions and util class methods in pacu.py
required a database session, combining them into a single class made
it easy to refactor the ubiquitous 'database' parameter into an
instance attribute.

Storing the database session and all utils on the same object also
opened the door to a simplification of module setup, since all pacu.py
utility methods now no longer need to include the database connection
as an argument when called.