Revert requirements.txt urllib3 version to 1.22
The requests package at version 2.18.4 raises errors internally if
the version of urllib3 available to it is anything other than 1.22.
The security patch introduced in urllib3 v1.23 does not affect Pacu's
use of the requests package, so this reversion will be used until the
requests package is updated.

References: https://nvd.nist.gov/vuln/detail/CVE-2018-20060
https://github.com/requests/requests/blob/master/Pipfile.lock#L56