Add iam__pivot
This uses pmmapper to build a graph db of escalation routes and
gives you the option of which one to traverse.

This only checks for assume role but can be extended to check for
other escalations.

This version of the module is fairly basic and doesn't attempt to
integrate with pacu, which means it won't know about previously
found resources and pacu won't know about any resources pmapper
finds.

For a different take on this same module see
features/iam__pivot-integrated.