Minimal fix when enumerating NotResource
This just prevents Pacu from erroring out when it can likely complete
the rest of the polices. There's likely other code path's this will
happen on but I'm not able to test those at the moment, think we'll
just have to fix them as they come up, or until we can rewrite this
code.