Remove read only modules
This is primarily to understand how much code needs testing because
it makes changes to the environment. Really all these should be
deduplicated into one module however, it is possible to download
data from the API dynamically (AWSPX and aws-inventory do this).

This is just a guess, we may need to dig into these further to
determine if they make any changes to then environment.