Initial draft for Organizations Assume Role module.
Initial draft for Organizations assume role module. Ideally it would be good if I could create a key set or give the user an option to create key set in background but NOT switch to it. Might be a good idea if Pacu just had an assume_role module anyone could call with some optional flags to pass in for these kinds of use cases.